+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 IMPROVEMENTS
 
 * Added `detach_ports_before_destroy` argument for `compute_instance_v2` resource that allows to detach all instance ports prior trying to delete the instance [GH-866]
+* Added web-download import method to `openstack_images_image_v2` resource [GH-888]
 
 BUG FIXES